Refactors the project, updates ownership
The previous project root had a bin folder that was used for a cheap way to have more than one program file in the project. After the project was properly refactored using modules, the binary folder was causing issues with IntelliJ's build index, so it was deleted.

This commit also adds examples to illustrate ownership and borrowing rules for heap-allocated types such as String and Vector.